+++ src/api/getContext.ts
@@ -6,7 +6,7 @@ import requireJson from '../fs/requireJson'
 import writePkg = require('write-pkg')
 import expandTilde, {isHomepath} from '../fs/expandTilde'
 import {StrictPnpmOptions} from '../types'
-import initLogger from '../logger'
+import initReporter from '../reporter'
 import {
   read as readGraph,
   Graph,
@@ -79,7 +79,7 @@ export default async function (opts: StrictPnpmOptions): Promise<PnpmContext> {
     isFirstInstallation,
   }
 
-  if (!opts.silent) initLogger(opts.logger)
+  if (!opts.silent) initReporter(opts.reporter)
 
   await mkdirp(ctx.cache)
   await mkdirp(ctx.storePath)
