+++ src/logging/logInstallStatus.ts
