+++ src/logging/streamParser.ts
