+++ src/reporter/index.ts
@@ -2,9 +2,9 @@ import supportsColor = require('supports-color')
 import pretty from './pretty'
 import simple from './simple'
 
-export type LoggerType = 'pretty' | 'simple'
+export type ReporterType = 'pretty' | 'simple'
 
-export default (loggerType: LoggerType) =>
-  ((loggerType === 'pretty' && supportsColor)
+export default (reporterType: ReporterType) =>
+  ((reporterType === 'pretty' && supportsColor)
     ? pretty()
     : simple())
