+++ tsconfig.json
@@ -46,7 +46,6 @@
     "src/cmd/uninstall.ts",
     "src/cmd/update.ts",
     "src/err.ts",
-    "src/fileLogger.ts",
     "src/fs/expandTilde.ts",
     "src/fs/getPkgDirs.ts",
     "src/fs/graphController.ts",
@@ -67,15 +66,18 @@
     "src/install/linkBins.ts",
     "src/install/linkPeers.ts",
     "src/install/postInstall.ts",
-    "src/logger/index.ts",
-    "src/logger/pretty.ts",
-    "src/logger/simple.ts",
     "src/memoize.ts",
+    "src/logging/fileLogger.ts",
+    "src/logging/logInstallStatus.ts",
+    "src/logging/streamParser.ts",
     "src/network/createCache.ts",
     "src/network/got.ts",
     "src/pnpmPkgJson.ts",
     "src/registryFor.ts",
     "src/removeDeps.ts",
+    "src/reporter/index.ts",
+    "src/reporter/pretty.ts",
+    "src/reporter/simple.ts",
     "src/resolve/createPkgId.ts",
     "src/resolve/fetch.ts",
     "src/resolve/getTarballName.ts",
